requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

aniso8601==8.0.0
apscheduler==3.6.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
flask-accepts==0.16.2
flask-cors==3.0.8
flask-restx==0.1.0
flask==1.1.1
idna==2.9
influxdb==5.2.3
itsdangerous==1.1.0
jinja2==3.0.0a1
jsonschema==3.2.0
markupsafe==1.1.1
marshmallow==3.5.1
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.4
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
pyzmq==19.0.0
requests==2.23.0
six==1.14.0
tzlocal==2.1b1
urllib3==1.25.8
werkzeug==0.16.0
